Move api fixture to api tests
